LibC: Conditionally include `<strings.h>`
Certain C Libraries have (unfortunately) included strings.h as a
part of string.h, which violates the POSIX spec for that specific
header. Some applications rely on this being the case. Linux uses
a `__USE_MISC` macro, which is always enabled by default, however
this macro specifically should be enabled per software package so
that we can maintain software compatibility, but not break POSIX
correctness.